+++ aur_scanner_gui/__init__.py
@@ -1,4 +1,4 @@
 """
 aur-scanner-gui package
 """
-__version__ = "1.2.0"
+__version__ = "1.2.5"